Cleanup postprocess scripts
- Add unique name to crawlers for better distinction
- Remove unit test leftovers
- Remove wrong DNS hierarchy script
- Fix log file name of url2hostname script
