+++ iyp/post/url2hostname.py
@@ -1,9 +1,13 @@
+import argparse
 import logging
+import os
 import sys
 
 import tldextract
 
 from iyp import BasePostProcess
+
+NAME = 'post.url2hostname'
 
 
 class PostProcess(BasePostProcess):
@@ -35,21 +39,35 @@ class PostProcess(BasePostProcess):
         # push links to IYP
         self.iyp.batch_add_links('PART_OF', links)
 
+    def unit_test(self):
+        raise NotImplementedError()
 
-if __name__ == '__main__':
 
-    scriptname = sys.argv[0].replace('/', '_')[0:-3]
-    FORMAT = '%(asctime)s %(processName)s %(message)s'
+def main() -> None:
+    parser = argparse.ArgumentParser()
+    parser.add_argument('--unit-test', action='store_true')
+    args = parser.parse_args()
+
+    scriptname = os.path.basename(sys.argv[0]).replace('/', '_')[0:-3]
+    FORMAT = '%(asctime)s %(levelname)s %(message)s'
     logging.basicConfig(
         format=FORMAT,
         filename='log/' + scriptname + '.log',
         level=logging.INFO,
         datefmt='%Y-%m-%d %H:%M:%S'
     )
-    logging.info('Start: %s' % sys.argv)
 
-    post = PostProcess()
-    post.run()
-    post.close()
+    logging.info(f'Started: {sys.argv}')
 
-    logging.info('End: %s' % sys.argv)
+    post = PostProcess(NAME)
+    if args.unit_test:
+        post.unit_test()
+    else:
+        post.run()
+        post.close()
+    logging.info(f'Finished: {sys.argv}')
+
+
+if __name__ == '__main__':
+    main()
+    sys.exit(0)
